+++ src/PommaLabs.KVLite.SQLite/PersistentCache.cs
@@ -21,14 +21,13 @@
 // DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT
 // OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-using CodeProject.ObjectPool.Specialized;
 using Microsoft.Data.Sqlite;
 using NodaTime;
 using PommaLabs.KVLite.Database;
 using PommaLabs.KVLite.Extensibility;
 using PommaLabs.KVLite.Resources;
-using PommaLabs.Thrower.Goodies;
-using PommaLabs.Thrower.Logging;
+using PommaLabs.KVLite.Thrower.Goodies;
+using PommaLabs.KVLite.Logging;
 using System;
 using System.IO;
 
@@ -63,10 +62,9 @@ namespace PommaLabs.KVLite.SQLite
         /// <param name="serializer">The serializer.</param>
         /// <param name="compressor">The compressor.</param>
         /// <param name="clock">The clock.</param>
-        /// <param name="memoryStreamPool">The memory stream pool.</param>
         /// <param name="random">The random number generator.</param>
-        public PersistentCache(PersistentCacheSettings settings, ISerializer serializer = null, ICompressor compressor = null, IClock clock = null, IMemoryStreamPool memoryStreamPool = null, IRandom random = null)
-            : base(settings, new SQLiteCacheConnectionFactory<PersistentCacheSettings>(settings, "ReadWriteCreate", "WAL"), serializer, compressor, clock, memoryStreamPool, random)
+        public PersistentCache(PersistentCacheSettings settings, ISerializer serializer = null, ICompressor compressor = null, IClock clock = null, IRandom random = null)
+            : base(settings, new SQLiteCacheConnectionFactory<PersistentCacheSettings>(settings, "ReadWriteCreate", "WAL"), serializer, compressor, clock, random)
         {
             // Connection string must be customized by each cache.
             UpdateConnectionString();