+++ src/PommaLabs.KVLite.WebApi/AbstractCacheController.cs
@@ -24,7 +24,7 @@
 using NodaTime;
 using NodaTime.Extensions;
 using PommaLabs.KVLite.Resources;
-using PommaLabs.Thrower;
+using PommaLabs.KVLite.Thrower;
 using System;
 using System.Collections.Generic;
 using System.Linq;