+++ src/PommaLabs.KVLite/AbstractAsyncCache.cs
@@ -23,8 +23,8 @@
 
 using NodaTime;
 using PommaLabs.KVLite.Resources;
-using PommaLabs.Thrower;
-using PommaLabs.Thrower.Logging;
+using PommaLabs.KVLite.Thrower;
+using PommaLabs.KVLite.Logging;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;