+++ test/PommaLabs.KVLite.UnitTests/NinjectConfig.cs
@@ -21,7 +21,6 @@
 // DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT
 // OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-using CodeProject.ObjectPool.Specialized;
 using Ninject.Modules;
 using NodaTime;
 using NodaTime.Testing;
@@ -37,10 +36,6 @@ namespace PommaLabs.KVLite.UnitTests
     {
         public override void Load()
         {
-            Bind<IMemoryStreamPool>()
-                .ToConstant(MemoryStreamPool.Instance)
-                .InSingletonScope();
-
             Bind<IClock>()
                 .ToConstant(new FakeClock(SystemClock.Instance.GetCurrentInstant()))
                 .InSingletonScope();